Add rake check and GH actions
